requirements.txt
flask
python-dotenv
langchain-openai
cdp-langchain
langgraph
langchain-core